+++ Under/DynamicFormula.h
@@ -11,44 +11,44 @@ using namespace Eigen;
 class DynamicFormula
 {
 public:
-	DynamicFormula(Vector3d omega, Vector3d velocity, Matrix3d R,
-		Vector3d y, Vector3d ts, Vector3d fs,MatrixXd K,double delta_t);
+	DynamicFormula(Vector3f omega, Vector3f velocity, Matrix3f R,
+		Vector3f y, Vector3f ts, Vector3f fs,MatrixXf K,float delta_t);
 	~DynamicFormula();
-	void setK(MatrixXd newK) {
+	void setK(MatrixXf newK) {
 		K = newK;
 	}
-	void setR(Matrix3d newR) {
+	void setR(Matrix3f newR) {
 		R = newR;
 	}
-	Matrix3d toDaOmegaOrY(Vector3d omega);
-	VectorXd tsfs2tf(Matrix3d R, Matrix3d Y);
-	Matrix3d computeR_();
-	Vector3d computey_();
-	VectorXd computelp();
+	Matrix3f toDaOmegaOrY(Vector3f omega);
+	VectorXf tsfs2tf(Matrix3f R, Matrix3f Y);
+	Matrix3f computeR_();
+	Vector3f computey_();
+	VectorXf computelp();
 
-	VectorXd computelp_(VectorXd lp);
-	void computeNextR(Matrix3d R_);
-	void computeNexty(Vector3d y_);
-	VectorXd computeNextlp(VectorXd lp, VectorXd lp_);
-	void computeNextwv(VectorXd lp);
-	Vector3d vec62Vec31(VectorXd ab);
-	Vector3d vec62Vec32(VectorXd ab);
+	VectorXf computelp_(VectorXf lp);
+	void computeNextR(Matrix3f R_);
+	void computeNexty(Vector3f y_);
+	VectorXf computeNextlp(VectorXf lp, VectorXf lp_);
+	void computeNextwv(VectorXf lp);
+	Vector3f vec62Vec31(VectorXf ab);
+	Vector3f vec62Vec32(VectorXf ab);
 	void nextTime();
-	void set_tsfs(Vector3d ts,Vector3d fs);
+	void set_tsfs(Vector3f ts,Vector3f fs);
 
 //private:
 	//初始时刻的 6+1个量
-	Vector3d w;
-	Vector3d v;
-	Matrix3d R; //R初始化一个正交矩阵
-	Vector3d y;
-	Vector3d ts;
-	Vector3d fs;
+	Vector3f w;
+	Vector3f v;
+	Matrix3f R; //R初始化一个正交矩阵
+	Vector3f y;
+	Vector3f ts;
+	Vector3f fs;
 
-	MatrixXd K;
-	double delta_t;
+	MatrixXf K;
+	float delta_t;
 	//平移旋转需要的两个参数
-	Quaterniond q;
-	Vector3d temp_deltay;
+	Quaternionf q;
+	Vector3f temp_deltay;
 
 };
